Pulled in library changes from batch_runner branch
